+++ vehicle/OVMS.V3/components/vehicle_maxus_edeliver3/src/vehicle_med3.cpp
@@ -127,7 +127,7 @@ OvmsVehicleMaxed3::~OvmsVehicleMaxed3()
 
 // IncomingPollReply:
 
-void OvmsVehicleMaxed3::IncomingPollReply(canbus* bus, uint16_t type, uint16_t pid, uint8_t* data, uint8_t length, uint16_t mlremain)
+void OvmsVehicleMaxed3::IncomingPollReply(canbus* bus, uint16_t type, uint32_t pid, uint8_t* data, uint8_t length, uint16_t mlremain)
 {
   // init / fill rx buffer:
   if (m_poll_ml_frame == 0) {
