+++ vehicle/OVMS.V3/components/vehicle_mitsubishi/src/mi_can_poll.cpp
@@ -29,7 +29,7 @@ static const char *TAG = "v-mitsubishi";
 /**
  * Incoming poll reply messages
  */
-void OvmsVehicleMitsubishi::IncomingPollReply(canbus* bus, uint16_t type, uint16_t pid, uint8_t* data, uint8_t length, uint16_t mlremain)
+void OvmsVehicleMitsubishi::IncomingPollReply(canbus* bus, uint16_t type, uint32_t pid, uint8_t* data, uint8_t length, uint16_t mlremain)
 {
   //ESP_LOGW(TAG, "%03x TYPE:%x PID:%02x Data:%02x %02x %02x %02x %02x %02x %02x %02x LENG:%02x REM:%02x", m_poll_moduleid_low, type, pid, data[0], data[1], data[2], data[3], data[4], data[5], data[6], data[7], length, mlremain);
